[experimental] Add an option to use SM for copying data
On HGX 8 A100, using the GPU's DMA engine through cudaMemcpy*() doesn't
give the best results when each GPU is copying to multiple targets. All
the copies from a single process (such as the memcpy_gemm binary) are
serialized and that can lead to severe under-utilization when the target
of the copy is oversubscribed.

This change adds a new option (-use_cudaComputeCopy) that switches all
the GPU <-> GPU flows to using the GPU SMs to do the copying. That
allows multiple copies to/from a single GPU to interleave and get a
better link utilization. This has been only tested and experimented with
on HGX 8 A100 and can be considered a proof of concept. Notably SM
copies have a lower peak bandwidth than CE, the code as implemented can
reach ~250 GB/s unidirectional and ~230 GB/s bidirectional.

Some results on HGX 8 A100 for reference to illustrate the SM vs CE
differences.

Unidirectional and bidirectional with SM pull:
--flows="g0-g1-a0 g2-g3-a1 g3-g2-a2" --buffer_size_KiB 131072 -use_cudaComputeCopy
255.631 (253.015)	230.940 (228.533)	230.806 (228.454)

Oversubcription with SM vs CE:
--flows="g0-g1-a0 g2-g3-a1 g3-g2-a2" --buffer_size_KiB 131072 -use_cudaComputeCopy
90.449 (88.780)	90.449 (88.746)	86.337 (85.173)	135.673 (133.544)

--flows="g0-g1-a0 g2-g3-a1 g3-g2-a2" --buffer_size_KiB 131072
101.519 (99.940)	104.794 (99.928)	65.496 (66.296)	68.771 (66.329)

3 GPUs sending to 3 other GPUs:
--flows=g0-g8-a0 g0-g9-a1 g0-g10-a2 g1-g8-a4 g1-g9-a5 g1-g10-a6 g2-g8-a8 g2-g9-a9 g2-g10-a10 --buffer_size_KiB 131072 -use_cudaComputeCopy
82.149 (81.334)	82.708 (81.269)	82.149 (80.781)	0.000 (0.000)	82.708 (81.530)	82.708 (81.318)	82.149 (80.944)	0.000 (0.000)	81.591 (81.237)	83.267 (80.960)	81.032 (80.813)

--flows=g0-g8-a0 g0-g9-a1 g0-g10-a2 g1-g8-a4 g1-g9-a5 g1-g10-a6 g2-g8-a8 g2-g9-a9 g2-g10-a10 --buffer_size_KiB 131072
29.431 (29.805)	29.431 (29.805)	29.615 (29.805)	0.000 (0.000)	29.431 (29.735)	29.431 (29.753)	29.615 (29.753)	0.000 (0.000)	29.431 (29.753)	29.431 (29.753)	29.615 (29.753)

The src/cuda_compute_copy.cu is included in git and needs to be locally
built with nvcc like so:
 $ nvcc -arch sm_80 src/cuda_compute_copy.cu -o src/cuda_compute_copy.a -lib -O3
And then bazel build can pick it up through cc_import(). Obviously this
is not ideal, but adding native cuda compilation support in bazel was
not tackled in this change.